controller fixes

+++ week1/express/src/components/Users/index.js
@@ -32,6 +32,15 @@ async function userCreate(req, res) {
 
 async function userFind(req, res) {
     try {
+        const id = parseInt(req.query.id, 10);
+
+        if (typeof id !== 'number' || Number.isNaN(id)) {
+            return res.status(400).json({
+                error: 'Error#100: Wrong input params.',
+                details: 'Id is not a number',
+            });
+        }
+
         const users = await UsersService.findUser(req.query);
 
         return res.status(201).json({
@@ -47,6 +56,15 @@ async function userFind(req, res) {
 
 async function userDelete(req, res) {
     try {
+        const id = parseInt(req.body.id, 10);
+
+        if (typeof id !== 'number' || Number.isNaN(id)) {
+            return res.status(400).json({
+                error: 'Error#100: Wrong input params.',
+                details: 'Id is not a number',
+            });
+        }
+
         const users = await UsersService.deleteUser(req.body);
 
         return res.status(201).json({
@@ -62,6 +80,14 @@ async function userDelete(req, res) {
 
 async function userUpdate(req, res) {
     try {
+        const id = parseInt(req.body.id, 10);
+
+        if (typeof id !== 'number' || Number.isNaN(id)) {
+            return res.status(400).json({
+                error: 'Error#100: Wrong input params.',
+                details: 'Id is not a number',
+            });
+        }
         const users = await UsersService.updateUser(req.body);
 
         return res.status(201).json({
